Split OneTimeActionType and RetainableActionType into two files
